refactor(user_service): create new endpoint for registration
