+++ microservice/api/services/user_service/login.go
@@ -7,7 +7,6 @@ import (
 	"log"
 	"microservice/api/common"
 	"net/http"
-	"strconv"
 
 	"github.com/jmoiron/sqlx"
 	_ "github.com/lib/pq"
@@ -21,53 +20,6 @@ const (
 )
 
 var connectionString = fmt.Sprintf("host=%s port=%d user=%s dbname=%s sslmode=disable", host, port, username, dbname)
-
-// TODO: This should be a separate endpoint
-func register(db *sqlx.DB, w http.ResponseWriter, r *http.Request, username string, password string) {
-	salt := generateSalt(saltLength)
-
-	passwordHash := hashPw([]byte(password), salt)
-
-	_, err := getUserFromName(db, username)
-
-	if err == nil {
-		log.Printf("Error: %v", err)
-		common.TryWriteResponse(w, "Username is already in use")
-		return
-	}
-
-	log.Println("Created new password salt")
-
-	var userId int
-	err = db.QueryRow("INSERT INTO users(username, pw_hash, salt) VALUES($1, $2, $3) RETURNING id", username, string(passwordHash), string(salt)).Scan(&userId)
-
-	if err != nil {
-		log.Printf("Error: %v", err)
-		common.TryWriteResponse(w, "Failed to create account")
-		return
-	}
-
-	log.Printf("Created new user")
-
-	if err != nil {
-		log.Printf("Error: %v", err)
-		common.TryWriteResponse(w, "Failed to create account")
-		return
-	}
-
-	session, err := createSession(db, userId)
-
-	if err != nil {
-		log.Printf("Error: %v", err)
-		common.TryWriteResponse(w, "Failed to create account")
-		return
-	}
-
-	// Not needed, but better be explicit!
-	w.Header().Set("Content-Type", "text/plain; charset")
-
-	common.TryWriteResponse(w, strconv.Itoa(session.ID))
-}
 
 func login(db *sqlx.DB, w http.ResponseWriter, r *http.Request, username string, password string) {
 	w.WriteHeader(http.StatusOK)
@@ -131,24 +83,6 @@ func LoginHandler(w http.ResponseWriter, r *http.Request) {
 
 	password := passwordParam[0]
 
-	var wantsToRegister bool
-
-	wantsToRegisterParam, okWantsToRegister := r.URL.Query()["wantsToRegister"]
-
-	if !okWantsToRegister {
-		wantsToRegister = false
-	} else {
-		var err error
-
-		wantsToRegister, err = strconv.ParseBool(wantsToRegisterParam[0])
-
-		if err != nil {
-			log.Printf("Error: %v", err)
-			common.TryWriteResponse(w, "Invalid value for parameter 'wantsToRegister', needs to be 'true' or 'false'")
-			return
-		}
-	}
-
 	db, err := sqlx.Open("postgres", connectionString)
 
 	defer db.Close()
@@ -161,11 +95,6 @@ func LoginHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if !wantsToRegister {
-		log.Println("Received login request")
-		login(db, w, r, username, password)
-	} else {
-		log.Println("Received registration request")
-		register(db, w, r, username, password)
-	}
+	log.Println("Received registration request")
+	login(db, w, r, username, password)
 }
